pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "aegix"
version = "0.1.0"
dependencies = [
  "typer",
  "docker",
  "PyYAML",
  "openai"
]
requires-python = ">=3.10"
authors = [
  {name = "Tongtong Zhu", email = "ttongg.zhu@gmail.com"},
]
description = "A secure, observable sandbox runtime for agent tool execution."
readme = "README.md"
license = "Apache-2.0"

[tool.hatch.build.targets.wheel]
packages = ["aegix_agent", "aegix_core", "aegix_cli"]

[project.scripts]
aegix = "aegix_core.cli:app"

